Extract Http request logic from providers, clean up directory structure and add DI
