+++ jobs/Backend/Task/Core/Http/IHttpClient.cs
@@ -1,0 +1,11 @@
+﻿using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace ExchangeRateUpdater.Core.Http
+{
+    public interface IHttpClient
+    {
+        Task<Stream> GetStreamAsync(string url, CancellationToken cancellationToken);
+    }
+}
